% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDims.R
\name{plotDims}
\alias{plotDims}
\title{Plot PCA and t-SNE
Output object of reductDims is specified as the input}
\usage{
plotDims(out)
}
\arguments{
\item{out}{Output object of reductDims}
}
\value{
Two scatter plots (PCA and t-SNE)
}
\description{
Plot PCA and t-SNE
Output object of reductDims is specified as the input
}
\examples{
input <- matrix(runif(50*100), nrow=50, ncol=100)
out <- reductDims(input, dim_pca=10, dim_tsne=2)
tmp <- tempfile()
png(file=tmp, width=1000, height=500)
reductDims::plotDims(out)
dev.off()
}
