% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductDims.R
\name{reductDims}
\alias{reductDims}
\title{Performs PCA and t-SNE
The eigenvectors of PCA are specified as the input of t-SNE}
\usage{
reductDims(
  input,
  dim_pca = 10,
  dim_tsne = 2,
  type_pca = c("svd", "irlba"),
  verbose = TRUE
)
}
\arguments{
\item{input}{A two-dimensional matrix}

\item{dim_pca}{The number of lower dimension of PCA}

\item{dim_tsne}{The number of lower dimension of t-SNE}

\item{type_pca}{"svd" or "irlba" can be specified (Default: svd)}

\item{verbose}{Verbose option to show the message (Default: TRUE)}
}
\value{
A list containing the results of PCA and t-SNE
}
\description{
Performs PCA and t-SNE
The eigenvectors of PCA are specified as the input of t-SNE
}
\examples{
input <- matrix(runif(50*100), nrow=50, ncol=100)
reductDims(input, dim_pca=10, dim_tsne=2)
}
